Refactor Desktop list
